+++ Controllers/PhonesController.cs
@@ -9,6 +9,9 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using TelefonSatis.Data;
 using TelefonSatis.Models;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Extensions;
+
 
 namespace TelefonSatis.Controllers
 {
@@ -21,6 +24,25 @@ namespace TelefonSatis.Controllers
             _context = context;
             _evrimoment = evrimoment;
         }
+        public bool SessionCont()
+        {
+            string permission = "";
+
+            if (HttpContext != null)
+            {
+                if ((HttpContext.Session.GetString("Permission")) != null)
+                {
+                    permission = HttpContext.Session.GetString("Permission");
+                    if (permission == "Admin")
+                    {
+                        return true;
+                    }
+                }
+            }
+            return false;
+        }
+
+
         #region default
         // GET: Phones
         public async Task<IActionResult> Index()
@@ -51,6 +73,11 @@ namespace TelefonSatis.Controllers
         // GET: Phones/Create
         public IActionResult Create()
         {
+            bool session = SessionCont();
+            if (session == false)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             ViewData["BrandId"] = new SelectList(_context.Brands, "BrandId", "BrandId");
             return View();
         }
@@ -62,6 +89,11 @@ namespace TelefonSatis.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("PhoneId,Name,Color,Price,MinDesc,Images,Description,TotalScore,TotalPeople,Score,BrandId")] Phone phone)
         {
+            bool session = SessionCont();
+            if (session == false)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (ModelState.IsValid)
             {
                 _context.Add(phone);
@@ -75,6 +107,11 @@ namespace TelefonSatis.Controllers
         // GET: Phones/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            bool session = SessionCont();
+            if (session == false)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -96,6 +133,11 @@ namespace TelefonSatis.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("PhoneId,Name,Color,Price,MinDesc,Description,TotalScore,TotalPeople,Score,BrandId")] Phone phone)
         {
+            bool session = SessionCont();
+            if (session == false)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id != phone.PhoneId)
             {
                 return NotFound();
@@ -128,6 +170,11 @@ namespace TelefonSatis.Controllers
         // GET: Phones/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            bool session = SessionCont();
+            if (session == false)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -145,6 +192,11 @@ namespace TelefonSatis.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            bool session = SessionCont();
+            if (session == false)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             var phone = await _context.Phones.FindAsync(id);
             _context.Phones.Remove(phone);
             await _context.SaveChangesAsync();
@@ -163,6 +215,11 @@ namespace TelefonSatis.Controllers
         // GET: Phones/Edit/5
         public async Task<IActionResult> Image(int? id)
         {
+            bool session = SessionCont();
+            if (session == false)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -184,7 +241,12 @@ namespace TelefonSatis.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Image([Bind("PhoneId","ResimDosyasi")] Phone phone)
         {
-            
+            bool session = SessionCont();
+            if (session == false)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
 
             if (ModelState.IsValid)
             {
